进一步完善DpmSetter的提示,修复I18NAttr的一个判断错误,并提高其性能

+++ AutumnBox.CoreModules/Extensions/NoMatter/EInstallExtension.cs
@@ -36,7 +36,7 @@ namespace AutumnBox.CoreModules.Extensions.NoMatter
                         string target = Path.Combine(extensionPath, file.Name);
                         file.CopyTo(target,true);
                         copied++;
-                        ui.Progress = (Convert.ToDouble(copied) /  files.Length * 100);
+                        ui.Progress = (int)(Convert.ToDouble(copied) /  files.Length * 100);
                     }
                     app.RestartApp();
                 }
